+++ R/data.R
@@ -1,0 +1,8 @@
+#' OneKP metadata file
+#'
+#' The object stored here should be exactly the same as the object returned
+#' from \code{retrieve_onekp()}. It is stored here for convenience and to save
+#' time in examples (\code{retrieve_onekp} takes around 30 seconds to run).
+#'
+#' @format OneKP object
+"onekp"