+++ R/filter.R
@@ -7,9 +7,7 @@
 #' @return OneKP object
 #' @name filter
 #' @examples
-#' \dontrun{
-#' onekp <- retrieve_onekp()
-#' onekp
+#' data(onekp)
 #'
 #' # filter by 1KP ID
 #' filter_by_code(onekp, c('URDJ', 'ROAP'))
@@ -25,7 +23,6 @@
 #'
 #' # filter by clade NCBI taxon ID
 #' filter_by_clade(onekp, 3700)
-#' }
 NULL
 
 #' @rdname filter